added working download function to script.
It will get the certificate id from the domain you want to use from the certificates.json file and fetch the certificate from harica and save it like *.key, *.pem and *.crt files in your local directory.

I've also implemented persistence of the private key when we do a req, that one will also pop up above.

I did stuff to be able to have the same keys as secrets gets from SUCM. Should be able to use them as certificate configuration for Apache directly.